+++ server/src/main/java/org/opensearch/common/settings/IndexScopedSettings.java
@@ -278,6 +278,8 @@ public final class IndexScopedSettings extends AbstractScopedSettings {
                 // Settings for search replica
                 IndexMetadata.INDEX_NUMBER_OF_SEARCH_REPLICAS_SETTING,
 
+                // Settings for Auto Force Merge
+                IndexSettings.INDEX_AUTO_FORCE_MERGES_ENABLED,
                 // Setting for derived source feature
                 IndexSettings.INDEX_DERIVED_SOURCE_SETTING,